+++ midi_adapter_flex/usb_descriptors.c
@@ -144,7 +144,7 @@ uint16_t const* tud_descriptor_string_cb(uint8_t index, uint16_t langid)
 {
   (void) langid;
 
-  uint8_t chr_count;
+  size_t chr_count;
 
   if ( index == 0)
   {
@@ -171,7 +171,7 @@ uint16_t const* tud_descriptor_string_cb(uint8_t index, uint16_t langid)
   }
 
   // first byte is length (including header), second byte is string type
-  _desc_str[0] = (TUSB_DESC_STRING << 8 ) | (2*chr_count + 2);
+  _desc_str[0] = (TUSB_DESC_STRING << 8 ) | (uint8_t) (2*chr_count + 2);
 
   return _desc_str;
 }
